[#30] refactor: p1-experience 단일 파일을 스테이지/데이터/타입으로 분해
