+++ frontend/features/smart-mirror/components/shared/audio-wave.tsx
