+++ frontend/features/smart-mirror/components/shared/glass-panel.tsx
