+++ frontend/features/smart-mirror/components/shared/scanning-effect.tsx
